+++ HISTORY.md
@@ -1,7 +1,7 @@
 
 # Version 0.5 
 - Major refactoring
-- Added __init__ generator
+- Added __init__ generatorgs
 
 # Version 0.4 
  - Major refactoring